requirements.txt
selenium==3.141.0
PyMySQL==0.9.3
google-api-core==1.7.0
google-auth==1.6.2
google-cloud==0.34.0
google-cloud-bigquery==1.8.1
google-cloud-core==0.29.1
google-cloud-vision==0.35.1
google-resumable-media==0.3.1
googleapis-common-protos==1.5.5
mysqlclient==1.3.13
bs4
grpcio==1.29.0
grpcio-tools==1.29.0
Flask